+++ src/bssunfold/core/__init__.py
@@ -23,6 +23,7 @@ from .unfold_statreg import solve_statreg
 from .unfold_reconst import solve_reconst
 from .unfold_scipy_direct_method import solve_scipy_direct
 from .unfold_tsvd import solve_tsvd
+from .unfold_lanczos import solve_lanczos
 from .unfold_parametric2 import solve_parametric2
 from .unfold_fruit_like import solve_fruit_like
 from .unfold_hybrid_parametric import solve_hybrid_parametric
@@ -68,6 +69,7 @@ from .unfold_statreg import unfold_statreg
 from .unfold_reconst import unfold_reconst
 from .unfold_scipy_direct_method import unfold_scipy_direct_method
 from .unfold_tsvd import unfold_tsvd
+from .unfold_lanczos import unfold_lanczos
 from .unfold_parametric2 import unfold_parametric2
 from .unfold_smt import unfold_smt
 from .unfold_scip import unfold_scip
@@ -104,6 +106,7 @@ __all__ = [
     "solve_reconst",
     "solve_scipy_direct",
     "solve_tsvd",
+    "solve_lanczos",
     "solve_parametric2",
     "solve_fruit_like",
     "solve_hybrid_parametric",
@@ -142,6 +145,7 @@ __all__ = [
     "unfold_reconst",
     "unfold_scipy_direct_method",
     "unfold_tsvd",
+    "unfold_lanczos",
     "unfold_parametric2",
     "unfold_fruit_like",
     "unfold_hybrid_parametric",